Add files via upload
Rock vs mine prediction is a binary classification problem that uses machine learning to predict whether an object detected by a sonar is a rock or a mine. This is a critical task for submarines and other underwater vehicles, as mines can be deadly if they are not detected and avoided.