Add files via upload
An email spam classifier is a machine learning model that is used to classify emails as spam or not spam. Spam emails are unsolicited and unwanted emails that are typically sent in bulk for commercial purposes. They can also be dangerous, as they may contain malicious links or attachments that can infect your computer with malware.